requirements.txt
absl-py==0.9.0
astunparse==1.6.3
atari-py==0.2.6
backcall==0.1.0
bsuite==0.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
dm-acme==0.1.2
dm-control==0.0.300771433
dm-env==1.2
dm-haiku==0.0.1
dm-reverb-nightly==0.1.0.dev20200605
dm-sonnet==2.0.0
dm-tree==0.1.5
EasyProcess==0.3
frozendict==1.2
future==0.18.2
gast==0.3.3
glfw==1.11.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gym==0.10.11
h5py==2.10.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
jax==0.1.69
jaxlib==0.1.47
jedi==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
matplotlib==3.2.1
mizani==0.7.1
networkx==2.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
palettable==3.3.0
pandas==1.0.4
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipe==1.6.0
plotnine==0.7.0
portpicker==1.3.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.3.2
Pygments==2.6.1
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyVirtualDisplay==1.3.2
PyWavelets==1.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rlax==0.0.1
rsa==4.0
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
statsmodels==0.11.1
tabulate==0.8.7
tb-nightly==2.3.0a20200605
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020060501
tf-nightly==2.3.0.dev20200604
tfp-nightly==0.11.0.dev20200605
tifffile==2020.6.3
traitlets==4.3.3
trfl==1.1.0
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
